+++ src/client/kotlin/me/gurkz/superdupermod/client/SuperDuperDataGenerator.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2026 Gurkan
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package me.gurkz.superdupermod.client
+
+import me.gurkz.superdupermod.client.datagen.SuperDuperModelProvider
+import me.gurkz.superdupermod.client.datagen.SuperDuperRecipeProvider
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator
+
+object SuperDuperDataGenerator : DataGeneratorEntrypoint {
+    override fun onInitializeDataGenerator(fabricDataGenerator: FabricDataGenerator) {
+        val pack = fabricDataGenerator.createPack()
+
+        pack.addProvider(::SuperDuperModelProvider)
+        pack.addProvider(::SuperDuperRecipeProvider)
+    }
+}