+++ src/client/kotlin/me/gurkz/superdupermod/client/datagen/SuperDuperModelProvider.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2026 Gurkan
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package me.gurkz.superdupermod.client.datagen
+
+import me.gurkz.superdupermod.item.ModItems
+import net.fabricmc.fabric.api.client.datagen.v1.provider.FabricModelProvider
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput
+import net.minecraft.client.data.models.BlockModelGenerators
+import net.minecraft.client.data.models.ItemModelGenerators
+import net.minecraft.client.data.models.model.ModelTemplates
+
+class SuperDuperModelProvider(output: FabricDataOutput) : FabricModelProvider(output) {
+    override fun generateBlockStateModels(blockModelGenerator: BlockModelGenerators) {
+
+    }
+
+    override fun generateItemModels(itemModelGenerator: ItemModelGenerators) {
+        itemModelGenerator.generateFlatItem(ModItems.SILENCER_STICK, ModelTemplates.FLAT_ITEM)
+    }
+
+    override fun getName(): String {
+        return "SuperDuperModelProvider"
+    }
+}